(improvement)Removed c.info[4]
A character object (`good.archer`) contains a property `.info` which contains numerous information about the character.

One of these is `.info[4]` which is the character's name.

The character name is already informed at `.name`.

This makes a duplicate variable which is bad in many ways:

- When adding new characters, I have to add the information for the name 3 times (now I will only have to add it twice)
- The app takes up more code causing longer loading times and more storage space
- The app processes more information causing it to be slower

So because of these reasons, I removed the duplicate variable to solve these problems.

I had to change the places that referenced `.info[5]` to be what is now `info[5]` but going to be `.info[4]`