(improvement)Cleaned up weapons (#149)
- Removed `class="theSword"`
- Removed `url(img/sword.png)` from multiplayer css
- Changed id's of swords to "leftWeapon" and "rightWeapon"

This all helps contributors (and us) to more easily recognize which weapons they are for.

First of all, the aren't all swords, so weapon is a better name
Second, the bSword was actually the left side which is always A's. So this was wrong, and once we add other letters, it would be even more wrong.