(issue)Used relative urls with "/index.html"
- Made credits and story play button use relative URL's
- Added `/index.html` to all URL's (except for store and index which will be added in a different pr)

This resolves issue #178.

This allows the iOS app porting to take less work, meaning that I will not have to change the URL's to relative URL's and I will not have to add the `/index.html` every time I update the iOS app.

This also means that when you are using a computer and coding on your local files, the links will go directly to the page and not make you have to click the file.

Also, this can fix other things such as as someone downloading the game.