(issue)Redesigned landing page (#175)
* (issue)Redesigned landing page

I redesigned the landing page to have a new background (designed by @ethanchoo) and a new layout.

This improves the game because:
- A better looking background
- A better looking page
- A more responsive page
- Cleaner code (100% cleaned and js with comments)

* (improvement)Removed loading logo

The loading logo is what shows when loading the page but in this case, it looks like it's taking a long time to show the buttons. It also uses more code so I think it's better to remove for this page.